Move contents of domain package into data model
