+++ app/src/main/java/com/jshvarts/healthreads/data/model/Book.kt
@@ -1,4 +1,4 @@
-package com.jshvarts.healthreads.domain
+package com.jshvarts.healthreads.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
