+++ app/src/main/java/com/jshvarts/healthreads/data/persistence/BookDatabase.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.jshvarts.healthreads.domain.Book
+import com.jshvarts.healthreads.data.model.Book
 
 private const val DATABASE_VERSION = 1
