+++ app/src/main/java/com/jshvarts/healthreads/ui/booklist/BookListViewState.kt
@@ -1,6 +1,6 @@
 package com.jshvarts.healthreads.ui.booklist
 
-import com.jshvarts.healthreads.domain.Book
+import com.jshvarts.healthreads.data.model.Book
 import com.jshvarts.healthreads.ui.ErrorType
 
 sealed class BookListViewState {
